+++ src/outknn.jl
@@ -6,7 +6,7 @@ Compute a kNN distance-based outlierness.
 Keyword arguments:
 * `metric` : Metric used to compute the dustances. See function `getknn`.
 * `k` : Nb. nearest neighbors to consider.
-* `algo` : Function summarizing the `k` distances to the neighbours.
+* `algo` : Function summarizing the `k` distances to the neighbors.
 * `scal` : Boolean. If `true`, each column of `X` is scaled before computing the outlierness.
 
 For each observation (row of `X`), the outlierness is defined by the summary  (e.g. by median) of the distances