+++ docs/src/news.md
@@ -3,7 +3,7 @@
 ## *Version 0.8.9*
 
 - News 
-    - Function **outknn**: Compute a KNN distance-based outlierness.
+    - Function **outknn**: Compute a kNN distance-based outlierness.
 
 - Modifications
     - Code cleaning.